+++ lib/berkeley_library/util/files.rb
@@ -30,8 +30,7 @@ module BerkeleyLibrary
       #
       # @param obj [Object] the object that might be an IO
       def writer_like?(obj)
-        # TODO: is it possible/desirable to loosen this? how strict is libxml2?
-        obj.is_a?(IO) || obj.is_a?(StringIO)
+        obj.respond_to?(:write) && obj.respond_to?(:close)
       end
 
     end